using-python/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==1.3.3
redis==5.0.1
python-docx==1.1.0
pdfplumber==0.10.3
pydantic==2.5.0
requests==2.31.0
alembic==1.12.1